[CST-6761] fix captcha failed test & added missing java docs
