+++ dspace-server-webapp/src/test/java/org/dspace/app/rest/RegistrationRestRepositoryIT.java
@@ -9,6 +9,11 @@ package org.dspace.app.rest;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.mock;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
@@ -22,9 +27,12 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.lang3.StringUtils;
 import org.dspace.app.rest.matcher.RegistrationMatcher;
 import org.dspace.app.rest.model.RegistrationRest;
+import org.dspace.app.rest.repository.RegistrationRestRepository;
 import org.dspace.app.rest.test.AbstractControllerIntegrationTest;
+import org.dspace.eperson.InvalidReCaptchaException;
 import org.dspace.eperson.RegistrationData;
 import org.dspace.eperson.dao.RegistrationDataDAO;
+import org.dspace.eperson.service.CaptchaService;
 import org.dspace.services.ConfigurationService;
 import org.hamcrest.Matchers;
 import org.junit.Test;
@@ -33,10 +41,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 public class RegistrationRestRepositoryIT extends AbstractControllerIntegrationTest {
 
     @Autowired
-    protected RegistrationDataDAO registrationDataDAO;
-
+    private CaptchaService captchaService;
     @Autowired
-    protected ConfigurationService configurationService;
+    private RegistrationDataDAO registrationDataDAO;
+    @Autowired
+    private ConfigurationService configurationService;
+    @Autowired
+    private RegistrationRestRepository registrationRestRepository;
 
     @Test
     public void findByTokenTestExistingUserTest() throws Exception {
@@ -210,6 +221,82 @@ public class RegistrationRestRepositoryIT extends AbstractControllerIntegrationT
                    .content(mapper.writeValueAsBytes(registrationRest))
                    .contentType(contentType))
                    .andExpect(status().isForbidden());
+    }
+
+    @Test
+    public void registrationFlowWithValidCaptchaTokenTest() throws Exception {
+        configurationService.setProperty("registration.verification.enabled", "true");
+
+        String captchaToken = "123456";
+        String captchaToken1 = "12345676866";
+
+        CaptchaService captchaServiceMock = mock(CaptchaService.class);
+
+        registrationRestRepository.setCaptchaService(captchaServiceMock);
+
+        doThrow(new InvalidReCaptchaException("Invalid captcha token"))
+            .when(captchaServiceMock).processResponse(any(), any());
+
+        doNothing().when(captchaServiceMock).processResponse(eq(captchaToken), eq("register_email"));
+
+        List<RegistrationData> registrationDataList = registrationDataDAO.findAll(context, RegistrationData.class);
+        assertEquals(0, registrationDataList.size());
+
+        ObjectMapper mapper = new ObjectMapper();
+        RegistrationRest registrationRest = new RegistrationRest();
+        registrationRest.setEmail(eperson.getEmail());
+        try {
+            // will throw InvalidReCaptchaException because 'X-Recaptcha-Token' not equal captchaToken
+            getClient().perform(post("/api/eperson/registrations")
+                       .header("X-Recaptcha-Token", captchaToken1)
+                       .content(mapper.writeValueAsBytes(registrationRest))
+                       .contentType(contentType))
+                       .andExpect(status().isForbidden());
+
+            getClient().perform(post("/api/eperson/registrations")
+                       .header("X-Recaptcha-Token", captchaToken)
+                       .content(mapper.writeValueAsBytes(registrationRest))
+                       .contentType(contentType))
+                       .andExpect(status().isCreated());
+
+            registrationDataList = registrationDataDAO.findAll(context, RegistrationData.class);
+            assertEquals(1, registrationDataList.size());
+            assertTrue(StringUtils.equalsIgnoreCase(registrationDataList.get(0).getEmail(), eperson.getEmail()));
+
+            String newEmail = "newEPersonTest@gmail.com";
+            registrationRest.setEmail(newEmail);
+            getClient().perform(post("/api/eperson/registrations")
+                       .header("X-Recaptcha-Token", captchaToken)
+                       .content(mapper.writeValueAsBytes(registrationRest))
+                       .contentType(contentType))
+                       .andExpect(status().isCreated());
+
+            registrationDataList = registrationDataDAO.findAll(context, RegistrationData.class);
+            assertTrue(registrationDataList.size() == 2);
+            assertTrue(StringUtils.equalsIgnoreCase(registrationDataList.get(0).getEmail(), newEmail) ||
+                       StringUtils.equalsIgnoreCase(registrationDataList.get(1).getEmail(), newEmail));
+
+            configurationService.setProperty("user.registration", false);
+
+            newEmail = "newEPersonTestTwo@gmail.com";
+            registrationRest.setEmail(newEmail);
+            getClient().perform(post("/api/eperson/registrations")
+                       .header("X-Recaptcha-Token", captchaToken)
+                       .content(mapper.writeValueAsBytes(registrationRest))
+                       .contentType(contentType))
+                       .andExpect(status().is(HttpServletResponse.SC_UNAUTHORIZED));
+
+            assertEquals(2, registrationDataList.size());
+            assertTrue(!StringUtils.equalsIgnoreCase(registrationDataList.get(0).getEmail(), newEmail) &&
+                       !StringUtils.equalsIgnoreCase(registrationDataList.get(1).getEmail(), newEmail));
+        } finally {
+            registrationRestRepository.setCaptchaService(captchaService);
+            Iterator<RegistrationData> iterator = registrationDataList.iterator();
+            while (iterator.hasNext()) {
+                RegistrationData registrationData = iterator.next();
+                registrationDataDAO.delete(context, registrationData);
+            }
+        }
     }
 
 }
